requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

arch==5.6.0
beautifulsoup4==4.12.2
dash==2.9.3
matplotlib==3.7.1
numpy==1.24.3
pandas==2.0.1
plotly==5.14.1
prophet==1.1.4
python-dateutil==2.8.2
requests==2.31.0
scikit-learn==1.3.0
scipy==1.10.1
seaborn==0.12.2
statsmodels==0.14.0
